Fix more math.

+++ test/NumSharp.UnitTest/LinearAlgebra/np.dot.Test.cs
@@ -34,8 +34,7 @@ namespace NumSharp.UnitTest.LinearAlgebra
         {
             var x = np.arange(3);
             var y = np.arange(3, 6);
-
-            int nd3 = (int)np.dot(x, y);
+            int nd3 = np.dot(x, y);
             Assert.IsTrue(nd3 == 14);
         }
